Rename classes to make more sense
- s/engine/piston/g
- s/runner/engine/g

Fixes #11
