+++ automation_engine/__init__.py
@@ -4,12 +4,12 @@ Automation Engine.
 A lightweight and flexible framework to automate things with MQTT.
 """
 
-from .engine import Engine
-from .runner import EngineRunner
+from .engine import AutomationEngine
+from .piston import Piston
 from .version import __version__
 
 __all__ = [
     "__version__",
-    "Engine",
-    "EngineRunner",
+    "Piston",
+    "AutomationEngine",
 ]
